Replace Gulp by Webpack
