+++ Resources/public/js/cmf_tree_browser.fancytree.js
@@ -1,0 +1,2 @@
+!function(e){function t(r){if(n[r])return n[r].exports;var i=n[r]={exports:{},id:r,loaded:!1};return e[r].call(i.exports,i,i.exports,t),i.loaded=!0,i.exports}var n={};return t.m=e,t.c=n,t.p="",t(0)}([function(e,t,n){"use strict";function r(e){return e&&e.__esModule?e:{"default":e}}var i=n(27),o=r(i),s=n(29);o["default"].fn.cmfTree=function(e){e=o["default"].extend({adapter:null,request:{load:null},actions:{}},e);var t=(0,o["default"])(this),n=function(e){if("string"==typeof e)return(0,o["default"])(e);if(e instanceof o["default"])return e;throw"Cannot handle selector "+e+". You may want to pass a jQuery object or a jQuery selector."};if(!e.request.load)throw"cmfTree needs an AJAX URL to lazy load the tree, pass it using the `request.load` option.";e.adapter||(e.adapter=new s.FancytreeAdapter(e));var r=e.adapter;if(!r.bindToElement)throw"cmfTree adapters must have a bindToElement() method to specify the output element of the tree.";for(var i in e.actions)if(e.actions.hasOwnProperty(i)){if(!r.addAction)throw"The configured cmfTree adapter does not support actions, implement the addAction() method or use another adapter.";var a=e.actions[i];if(!a.url)throw'actions should have a url defined, "'+i+'" does not.';r.addAction(i,a.url,a.icon)}if(r.bindToElement(t),e.path_output){if(!r.bindToInput)throw"The configured cmfTree adapter does not support binding to an input field, implement the bindToInput() method or use another adapter.";r.bindToInput(n(e.path_output))}return r}},function(e,t,n){var r=n(43)("wks"),i=n(15),o=n(2).Symbol;e.exports=function(e){return r[e]||(r[e]=o&&o[e]||(o||i)("Symbol."+e))}},function(e,t){var n=e.exports="undefined"!=typeof window&&window.Math==Math?window:"undefined"!=typeof self&&self.Math==Math?self:Function("return this")();"number"==typeof __g&&(__g=n)},function(e,t,n){var r=n(4),i=n(23);e.exports=n(10)?function(e,t,n){return r.setDesc(e,t,i(1,n))}:function(e,t,n){return e[t]=n,e}},function(e,t){var n=Object;e.exports={create:n.create,getProto:n.getPrototypeOf,isEnum:{}.propertyIsEnumerable,getDesc:n.getOwnPropertyDescriptor,setDesc:n.defineProperty,setDescs:n.defineProperties,getKeys:n.keys,getNames:n.getOwnPropertyNames,getSymbols:n.getOwnPropertySymbols,each:[].forEach}},function(e,t){e.exports={}},function(e,t,n){var r=n(2),i=n(3),o=n(15)("src"),s="toString",a=Function[s],l=(""+a).split(s);n(7).inspectSource=function(e){return a.call(e)},(e.exports=function(e,t,n,s){"function"==typeof n&&(n.hasOwnProperty(o)||i(n,o,e[t]?""+e[t]:l.join(String(t))),n.hasOwnProperty("name")||i(n,"name",t)),e===r?e[t]=n:(s||delete e[t],i(e,t,n))})(Function.prototype,s,function(){return"function"==typeof this&&this[o]||a.call(this)})},function(e,t){var n=e.exports={version:"1.2.6"};"number"==typeof __e&&(__e=n)},function(e,t,n){var r=n(32);e.exports=function(e,t,n){if(r(e),void 0===t)return e;switch(n){case 1:return function(n){return e.call(t,n)};case 2:return function(n,r){return e.call(t,n,r)};case 3:return function(n,r,i){return e.call(t,n,r,i)}}return function(){return e.apply(t,arguments)}}},function(e,t){e.exports=function(e){if(void 0==e)throw TypeError("Can't call method on  "+e);return e}},function(e,t,n){e.exports=!n(20)(function(){return 7!=Object.defineProperty({},"a",{get:function(){return 7}}).a})},function(e,t){var n={}.hasOwnProperty;e.exports=function(e,t){return n.call(e,t)}},function(e,t){e.exports=function(e){return"object"==typeof e?null!==e:"function"==typeof e}},function(e,t,n){"use strict";var r=n(41),i=n(19),o=n(6),s=n(3),a=n(11),l=n(5),c=n(39),u=n(14),d=n(4).getProto,h=n(1)("iterator"),f=!([].keys&&"next"in[].keys()),p="@@iterator",v="keys",g="values",y=function(){return this};e.exports=function(e,t,n,m,b,x,_){c(n,t,m);var k,N,w=function(e){if(!f&&e in T)return T[e];switch(e){case v:return function(){return new n(this,e)};case g:return function(){return new n(this,e)}}return function(){return new n(this,e)}},S=t+" Iterator",C=b==g,E=!1,T=e.prototype,A=T[h]||T[p]||b&&T[b],L=A||w(b);if(A){var P=d(L.call(new e));u(P,S,!0),!r&&a(T,p)&&s(P,h,y),C&&A.name!==g&&(E=!0,L=function(){return A.call(this)})}if(r&&!_||!f&&!E&&T[h]||s(T,h,L),l[t]=L,l[S]=y,b)if(k={values:C?L:w(g),keys:x?L:w(v),entries:C?w("entries"):L},_)for(N in k)N in T||o(T,N,k[N]);else i(i.P+i.F*(f||E),t,k);return k}},function(e,t,n){var r=n(4).setDesc,i=n(11),o=n(1)("toStringTag");e.exports=function(e,t,n){e&&!i(e=n?e:e.prototype,o)&&r(e,o,{configurable:!0,value:t})}},function(e,t){var n=0,r=Math.random();e.exports=function(e){return"Symbol(".concat(void 0===e?"":e,")_",(++n+r).toString(36))}},function(e,t,n){var r=n(12);e.exports=function(e){if(!r(e))throw TypeError(e+" is not an object!");return e}},function(e,t,n){var r=n(18),i=n(1)("toStringTag"),o="Arguments"==r(function(){return arguments}());e.exports=function(e){var t,n,s;return void 0===e?"Undefined":null===e?"Null":"string"==typeof(n=(t=Object(e))[i])?n:o?r(t):"Object"==(s=r(t))&&"function"==typeof t.callee?"Arguments":s}},function(e,t){var n={}.toString;e.exports=function(e){return n.call(e).slice(8,-1)}},function(e,t,n){var r=n(2),i=n(7),o=n(3),s=n(6),a=n(8),l="prototype",c=function(e,t,n){var u,d,h,f,p=e&c.F,v=e&c.G,g=e&c.S,y=e&c.P,m=e&c.B,b=v?r:g?r[t]||(r[t]={}):(r[t]||{})[l],x=v?i:i[t]||(i[t]={}),_=x[l]||(x[l]={});v&&(n=t);for(u in n)d=!p&&b&&u in b,h=(d?b:n)[u],f=m&&d?a(h,r):y&&"function"==typeof h?a(Function.call,h):h,b&&!d&&s(b,u,h),x[u]!=h&&o(x,u,f),y&&_[u]!=h&&(_[u]=h)};r.core=i,c.F=1,c.G=2,c.S=4,c.P=8,c.B=16,c.W=32,e.exports=c},function(e,t){e.exports=function(e){try{return!!e()}catch(t){return!0}}},function(e,t,n){var r=n(8),i=n(38),o=n(37),s=n(16),a=n(46),l=n(47);e.exports=function(e,t,n,c){var u,d,h,f=l(e),p=r(n,c,t?2:1),v=0;if("function"!=typeof f)throw TypeError(e+" is not iterable!");if(o(f))for(u=a(e.length);u>v;v++)t?p(s(d=e[v])[0],d[1]):p(e[v]);else for(h=f.call(e);!(d=h.next()).done;)i(h,p,d.value,t)}},function(e,t){e.exports=function(e,t){return{value:t,done:!!e}}},function(e,t){e.exports=function(e,t){return{enumerable:!(1&e),configurable:!(2&e),writable:!(4&e),value:t}}},function(e,t,n){var r=n(6);e.exports=function(e,t){for(var n in t)r(e,n,t[n]);return e}},function(e,t){e.exports=function(e,t,n){if(!(e instanceof t))throw TypeError(n+": use the 'new' operator!");return e}},function(e,t){var n=Math.ceil,r=Math.floor;e.exports=function(e){return isNaN(e=+e)?0:(e>0?r:n)(e)}},function(e,t){e.exports=jQuery},function(e,t){!function(e,t,n,r){"use strict";function i(t,n){t||(n=n?": "+n:"",e.error("Fancytree assertion failed"+n))}function o(e,n){var r,i,o=t.console?t.console[e]:null;if(o)try{o.apply(t.console,n)}catch(s){for(i="",r=0;r<n.length;r++)i+=n[r];o(i)}}function s(e){return!(!e.tree||e.statusNodeType===r)}function a(t,n,r,i){var o,s,a,l=e.map(e.trim(t).split("."),function(e){return parseInt(e,10)}),c=e.map(Array.prototype.slice.call(arguments,1),function(e){return parseInt(e,10)});for(o=0;o<c.length;o++)if(s=l[o]||0,a=c[o]||0,s!==a)return s>a;return!0}function l(e,t,n,r,i){var o=function(){var n=t[e],o=r[e],s=t.ext[i],a=function(){return n.apply(t,arguments)};return function(){var e=t._local,n=t._super;try{return t._local=s,t._super=a,o.apply(t,arguments)}finally{t._local=e,t._super=n}}}();return o}function c(t,n,r,i){for(var o in r)"function"==typeof r[o]?"function"==typeof t[o]?t[o]=l(o,t,n,r,i):"_"===o.charAt(0)?t.ext[i][o]=l(o,t,n,r,i):e.error("Could not override tree."+o+". Use prefix '_' to create tree."+i+"._"+o):"options"!==o&&(t.ext[i][o]=r[o])}function u(t,n){return t===r?e.Deferred(function(){this.resolve()}).promise():e.Deferred(function(){this.resolveWith(t,n)}).promise()}function d(t,n){return t===r?e.Deferred(function(){this.reject()}).promise():e.Deferred(function(){this.rejectWith(t,n)}).promise()}function h(e,t){return function(){e.resolveWith(t)}}function f(t){var n=e.extend({},t.data()),r=n.json;return delete n.fancytree,r&&(delete n.json,n=e.extend(n,r)),n}function p(e){return e=e.toLowerCase(),function(t){return t.title.toLowerCase().indexOf(e)>=0}}function v(e){var t=new RegExp("^"+e,"i");return function(e){return t.test(e.title)}}function g(t,n){var r,o,s,a;for(this.parent=t,this.tree=t.tree,this.ul=null,this.li=null,this.statusNodeType=null,this._isLoading=!1,this._error=null,this.data={},r=0,o=S.length;o>r;r++)s=S[r],this[s]=n[s];n.data&&e.extend(this.data,n.data);for(s in n)C[s]||e.isFunction(n[s])||E[s]||(this.data[s]=n[s]);null==this.key?this.tree.options.defaultKey?(this.key=this.tree.options.defaultKey(this),i(this.key,"defaultKey() must return a unique key")):this.key="_"+b._nextNodeKey++:this.key=""+this.key,n.active&&(i(null===this.tree.activeNode,"only one active node allowed"),this.tree.activeNode=this),n.selected&&(this.tree.lastSelectedNode=this),this.children=null,a=n.children,a&&a.length&&this._setChildren(a),this.tree._callHook("treeRegisterNode",this.tree,!0,this)}function y(t){this.widget=t,this.$div=t.element,this.options=t.options,this.options&&(e.isFunction(this.options.lazyload)&&!e.isFunction(this.options.lazyLoad)&&(this.options.lazyLoad=function(){return b.warn("The 'lazyload' event is deprecated since 2014-02-25. Use 'lazyLoad' (with uppercase L) instead."),t.options.lazyload.apply(this,arguments)}),e.isFunction(this.options.loaderror)&&e.error("The 'loaderror' event was renamed since 2014-07-03. Use 'loadError' (with uppercase E) instead."),this.options.fx!==r&&b.warn("The 'fx' options was replaced by 'toggleEffect' since 2014-11-30.")),this.ext={},this.data=f(this.$div),this._id=e.ui.fancytree._nextId++,this._ns=".fancytree-"+this._id,this.activeNode=null,this.focusNode=null,this._hasFocus=null,this.lastSelectedNode=null,this.systemFocusElement=null,this.lastQuicksearchTerm="",this.lastQuicksearchTime=0,this.statusClassPropName="span",this.ariaPropName="li",this.nodeContainerAttrName="li",this.$div.find(">ul.fancytree-container").remove();var n,i={tree:this};this.rootNode=new g(i,{title:"root",key:"root_"+this._id,children:null,expanded:!0}),this.rootNode.parent=null,n=e("<ul>",{"class":"ui-fancytree fancytree-container"}).appendTo(this.$div),this.$container=n,this.rootNode.ul=n[0],null==this.options.debugLevel&&(this.options.debugLevel=b.debugLevel),this.$container.attr("tabindex",this.options.tabbable?"0":"-1"),this.options.aria&&this.$container.attr("role","tree").attr("aria-multiselectable",!0)}if(e.ui&&e.ui.fancytree)return void e.ui.fancytree.warn("Fancytree: ignored duplicate include");i(e.ui,"Fancytree requires jQuery UI (http://jqueryui.com)");var m,b=null,x={"&":"&amp;","<":"&lt;",">":"&gt;",'"':"&quot;","'":"&#39;","/":"&#x2F;"},_={16:!0,17:!0,18:!0},k={8:"backspace",9:"tab",10:"return",13:"return",19:"pause",20:"capslock",27:"esc",32:"space",33:"pageup",34:"pagedown",35:"end",36:"home",37:"left",38:"up",39:"right",40:"down",45:"insert",46:"del",59:";",61:"=",96:"0",97:"1",98:"2",99:"3",100:"4",101:"5",102:"6",103:"7",104:"8",105:"9",106:"*",107:"+",109:"-",110:".",111:"/",112:"f1",113:"f2",114:"f3",115:"f4",116:"f5",117:"f6",118:"f7",119:"f8",120:"f9",121:"f10",122:"f11",123:"f12",144:"numlock",145:"scroll",173:"-",186:";",187:"=",188:",",189:"-",190:".",191:"/",192:"`",219:"[",220:"\\",221:"]",222:"'"},N="active expanded focus folder hideCheckbox lazy selected unselectable".split(" "),w={},S="expanded extraClasses folder hideCheckbox key lazy refKey selected title tooltip unselectable".split(" "),C={},E={active:!0,children:!0,data:!0,focus:!0};for(m=0;m<N.length;m++)w[N[m]]=!0;for(m=0;m<S.length;m++)C[S[m]]=!0;g.prototype={_findDirectChild:function(e){var t,n,r=this.children;if(r)if("string"==typeof e){for(t=0,n=r.length;n>t;t++)if(r[t].key===e)return r[t]}else{if("number"==typeof e)return this.children[e];if(e.parent===this)return e}return null},_setChildren:function(e){i(e&&(!this.children||0===this.children.length),"only init supported"),this.children=[];for(var t=0,n=e.length;n>t;t++)this.children.push(new g(this,e[t]))},addChildren:function(t,n){var r,o,s,a=null,l=[];for(e.isPlainObject(t)&&(t=[t]),this.children||(this.children=[]),r=0,o=t.length;o>r;r++)l.push(new g(this,t[r]));return a=l[0],null==n?this.children=this.children.concat(l):(n=this._findDirectChild(n),s=e.inArray(n,this.children),i(s>=0,"insertBefore must be an existing child"),this.children.splice.apply(this.children,[s,0].concat(l))),(!this.parent||this.parent.ul||this.tr)&&this.render(),3===this.tree.options.selectMode&&this.fixSelection3FromEndNodes(),a},addNode:function(e,t){switch((t===r||"over"===t)&&(t="child"),t){case"after":return this.getParent().addChildren(e,this.getNextSibling());case"before":return this.getParent().addChildren(e,this);case"firstChild":var n=this.children?this.children[0]:null;return this.addChildren(e,n);case"child":case"over":return this.addChildren(e)}i(!1,"Invalid mode: "+t)},appendSibling:function(e){return this.addNode(e,"after")},applyPatch:function(t){if(null===t)return this.remove(),u(this);var n,r,i,o={children:!0,expanded:!0,parent:!0};for(n in t)i=t[n],o[n]||e.isFunction(i)||(C[n]?this[n]=i:this.data[n]=i);return t.hasOwnProperty("children")&&(this.removeChildren(),t.children&&this._setChildren(t.children)),this.isVisible()&&(this.renderTitle(),this.renderStatus()),r=t.hasOwnProperty("expanded")?this.setExpanded(t.expanded):u(this)},collapseSiblings:function(){return this.tree._callHook("nodeCollapseSiblings",this)},copyTo:function(e,t,n){return e.addNode(this.toDict(!0,n),t)},countChildren:function(e){var t,n,r,i=this.children;if(!i)return 0;if(r=i.length,e!==!1)for(t=0,n=r;n>t;t++)r+=i[t].countChildren();return r},debug:function(e){this.tree.options.debugLevel>=2&&(Array.prototype.unshift.call(arguments,this.toString()),o("log",arguments))},discard:function(){return this.warn("FancytreeNode.discard() is deprecated since 2014-02-16. Use .resetLazy() instead."),this.resetLazy()},findAll:function(t){t=e.isFunction(t)?t:p(t);var n=[];return this.visit(function(e){t(e)&&n.push(e)}),n},findFirst:function(t){t=e.isFunction(t)?t:p(t);var n=null;return this.visit(function(e){return t(e)?(n=e,!1):void 0}),n},_changeSelectStatusAttrs:function(e){var t=!1;switch(e){case!1:t=this.selected||this.partsel,this.selected=!1,this.partsel=!1;break;case!0:t=!this.selected||!this.partsel,this.selected=!0,this.partsel=!0;break;case r:t=this.selected||!this.partsel,this.selected=!1,this.partsel=!0;break;default:i(!1,"invalid state: "+e)}return t&&this.renderStatus(),t},fixSelection3AfterClick:function(){var e=this.isSelected();this.visit(function(t){t._changeSelectStatusAttrs(e)}),this.fixSelection3FromEndNodes()},fixSelection3FromEndNodes:function(){function e(t){var n,i,o,s,a,l,c,u=t.children;if(u&&u.length){for(l=!0,c=!1,n=0,i=u.length;i>n;n++)o=u[n],s=e(o),s!==!1&&(c=!0),s!==!0&&(l=!1);a=l?!0:c?r:!1}else a=!!t.selected;return t._changeSelectStatusAttrs(a),a}i(3===this.tree.options.selectMode,"expected selectMode 3"),e(this),this.visitParents(function(e){var t,n,i,o,s=e.children,a=!0,l=!1;for(t=0,n=s.length;n>t;t++)i=s[t],(i.selected||i.partsel)&&(l=!0),i.unselectable||i.selected||(a=!1);o=a?!0:l?r:!1,e._changeSelectStatusAttrs(o)})},fromDict:function(t){for(var n in t)C[n]?this[n]=t[n]:"data"===n?e.extend(this.data,t.data):e.isFunction(t[n])||E[n]||(this.data[n]=t[n]);t.children&&(this.removeChildren(),this.addChildren(t.children)),this.renderTitle()},getChildren:function(){return this.hasChildren()===r?r:this.children},getFirstChild:function(){return this.children?this.children[0]:null},getIndex:function(){return e.inArray(this,this.parent.children)},getIndexHier:function(t){t=t||".";var n=[];return e.each(this.getParentList(!1,!0),function(e,t){n.push(t.getIndex()+1)}),n.join(t)},getKeyPath:function(e){var t=[],n=this.tree.options.keyPathSeparator;return this.visitParents(function(e){e.parent&&t.unshift(e.key)},!e),n+t.join(n)},getLastChild:function(){return this.children?this.children[this.children.length-1]:null},getLevel:function(){for(var e=0,t=this.parent;t;)e++,t=t.parent;return e},getNextSibling:function(){if(this.parent){var e,t,n=this.parent.children;for(e=0,t=n.length-1;t>e;e++)if(n[e]===this)return n[e+1]}return null},getParent:function(){return this.parent},getParentList:function(e,t){for(var n=[],r=t?this:this.parent;r;)(e||r.parent)&&n.unshift(r),r=r.parent;return n},getPrevSibling:function(){if(this.parent){var e,t,n=this.parent.children;for(e=1,t=n.length;t>e;e++)if(n[e]===this)return n[e-1]}return null},hasChildren:function(){return this.lazy?null==this.children?r:0===this.children.length?!1:1===this.children.length&&this.children[0].isStatusNode()?r:!0:!(!this.children||!this.children.length)},hasFocus:function(){return this.tree.hasFocus()&&this.tree.focusNode===this},info:function(e){this.tree.options.debugLevel>=1&&(Array.prototype.unshift.call(arguments,this.toString()),o("info",arguments))},isActive:function(){return this.tree.activeNode===this},isChildOf:function(e){return this.parent&&this.parent===e},isDescendantOf:function(e){if(!e||e.tree!==this.tree)return!1;for(var t=this.parent;t;){if(t===e)return!0;t=t.parent}return!1},isExpanded:function(){return!!this.expanded},isFirstSibling:function(){var e=this.parent;return!e||e.children[0]===this},isFolder:function(){return!!this.folder},isLastSibling:function(){var e=this.parent;return!e||e.children[e.children.length-1]===this},isLazy:function(){return!!this.lazy},isLoaded:function(){return!this.lazy||this.hasChildren()!==r},isLoading:function(){return!!this._isLoading},isRoot:function(){return this.isRootNode()},isRootNode:function(){return this.tree.rootNode===this},isSelected:function(){return!!this.selected},isStatusNode:function(){return!!this.statusNodeType},isTopLevel:function(){return this.tree.rootNode===this.parent},isUndefined:function(){return this.hasChildren()===r},isVisible:function(){var e,t,n=this.getParentList(!1,!1);for(e=0,t=n.length;t>e;e++)if(!n[e].expanded)return!1;return!0},lazyLoad:function(e){return this.warn("FancytreeNode.lazyLoad() is deprecated since 2014-02-16. Use .load() instead."),this.load(e)},load:function(e){var t,n,r=this;return i(this.isLazy(),"load() requires a lazy node"),e||this.isUndefined()?(this.isLoaded()&&this.resetLazy(),n=this.tree._triggerNodeEvent("lazyLoad",this),n===!1?u(this):(i("boolean"!=typeof n,"lazyLoad event must return source in data.result"),t=this.tree._callHook("nodeLoadChildren",this,n),this.expanded&&t.always(function(){r.render()}),t)):u(this)},makeVisible:function(t){var n,r=this,i=[],o=new e.Deferred,s=this.getParentList(!1,!1),a=s.length,l=!(t&&t.noAnimation===!0),c=!(t&&t.scrollIntoView===!1);for(n=a-1;n>=0;n--)i.push(s[n].setExpanded(!0,t));return e.when.apply(e,i).done(function(){c?r.scrollIntoView(l).done(function(){o.resolve()}):o.resolve()}),o.promise()},moveTo:function(t,n,o){(n===r||"over"===n)&&(n="child");var s,a=this.parent,l="child"===n?t:t.parent;if(this!==t){if(!this.parent)throw"Cannot move system root";if(l.isDescendantOf(this))throw"Cannot move a node to its own descendant";if(1===this.parent.children.length){if(this.parent===l)return;this.parent.children=this.parent.lazy?[]:null,this.parent.expanded=!1}else s=e.inArray(this,this.parent.children),i(s>=0),this.parent.children.splice(s,1);if(this.parent=l,l.hasChildren())switch(n){case"child":l.children.push(this);break;case"before":s=e.inArray(t,l.children),i(s>=0),l.children.splice(s,0,this);break;case"after":s=e.inArray(t,l.children),i(s>=0),l.children.splice(s+1,0,this);break;default:throw"Invalid mode "+n}else l.children=[this];o&&t.visit(o,!0),this.tree!==t.tree&&(this.warn("Cross-tree moveTo is experimantal!"),this.visit(function(e){e.tree=t.tree},!0)),a.isDescendantOf(l)||a.render(),l.isDescendantOf(a)||l===a||l.render()}},navigate:function(t,n){function r(r){if(r){try{r.makeVisible()}catch(i){}return e(r.span).is(":visible")?n===!1?r.setFocus():r.setActive():(r.debug("Navigate: skipping hidden node"),void r.navigate(t,n))}}var i,o,s=!0,a=e.ui.keyCode,l=null;switch(t){case a.BACKSPACE:this.parent&&this.parent.parent&&r(this.parent);break;case a.LEFT:this.expanded?(this.setExpanded(!1),r(this)):this.parent&&this.parent.parent&&r(this.parent);break;case a.RIGHT:this.expanded||!this.children&&!this.lazy?this.children&&this.children.length&&r(this.children[0]):(this.setExpanded(),r(this));break;case a.UP:for(l=this.getPrevSibling();l&&!e(l.span).is(":visible");)l=l.getPrevSibling();for(;l&&l.expanded&&l.children&&l.children.length;)l=l.children[l.children.length-1];!l&&this.parent&&this.parent.parent&&(l=this.parent),r(l);break;case a.DOWN:if(this.expanded&&this.children&&this.children.length)l=this.children[0];else for(o=this.getParentList(!1,!0),i=o.length-1;i>=0;i--){for(l=o[i].getNextSibling();l&&!e(l.span).is(":visible");)l=l.getNextSibling();if(l)break}r(l);break;default:s=!1}},remove:function(){return this.parent.removeChild(this)},removeChild:function(e){return this.tree._callHook("nodeRemoveChild",this,e)},removeChildren:function(){return this.tree._callHook("nodeRemoveChildren",this)},render:function(e,t){return this.tree._callHook("nodeRender",this,e,t)},renderTitle:function(){return this.tree._callHook("nodeRenderTitle",this)},renderStatus:function(){return this.tree._callHook("nodeRenderStatus",this)},resetLazy:function(){this.removeChildren(),this.expanded=!1,this.lazy=!0,this.children=r,this.renderStatus()},scheduleAction:function(e,t){this.tree.timer&&clearTimeout(this.tree.timer),this.tree.timer=null;var n=this;switch(e){case"cancel":break;case"expand":this.tree.timer=setTimeout(function(){n.tree.debug("setTimeout: trigger expand"),n.setExpanded(!0)},t);break;case"activate":this.tree.timer=setTimeout(function(){n.tree.debug("setTimeout: trigger activate"),n.setActive(!0)},t);break;default:throw"Invalid mode "+e}},scrollIntoView:function(o,a){a!==r&&s(a)&&(this.warn("scrollIntoView() with 'topNode' option is deprecated since 2014-05-08. Use 'options.topNode' instead."),a={topNode:a});var l,c,u,d,h=e.extend({effects:o===!0?{duration:200,queue:!1}:o,scrollOfs:this.tree.options.scrollOfs,scrollParent:this.tree.options.scrollParent||this.tree.$container,topNode:null},a),f=new e.Deferred,p=this,v=e(this.span).height(),g=e(h.scrollParent),y=h.scrollOfs.top||0,m=h.scrollOfs.bottom||0,b=g.height(),x=g.scrollTop(),_=g,k=g[0]===t,N=h.topNode||null,w=null;return i(e(this.span).is(":visible"),"scrollIntoView node is invisible"),k?(c=e(this.span).offset().top,l=N&&N.span?e(N.span).offset().top:0,_=e("html,body")):(i(g[0]!==n&&g[0]!==n.body,"scrollParent should be an simple element or `window`, not document or body."),d=g.offset().top,c=e(this.span).offset().top-d+x,l=N?e(N.span).offset().top-d+x:0,u=Math.max(0,g.innerHeight()-g[0].clientHeight),b-=u),x+y>c?w=c-y:c+v>x+b-m&&(w=c+v-b+m,N&&(i(N.isRoot()||e(N.span).is(":visible"),"topNode must be visible"),w>l&&(w=l-y))),null!==w?h.effects?(h.effects.complete=function(){f.resolveWith(p)},_.stop(!0).animate({scrollTop:w},h.effects)):(_[0].scrollTop=w,f.resolveWith(this)):f.resolveWith(this),f.promise()},setActive:function(e,t){return this.tree._callHook("nodeSetActive",this,e,t)},setExpanded:function(e,t){return this.tree._callHook("nodeSetExpanded",this,e,t)},setFocus:function(e){return this.tree._callHook("nodeSetFocus",this,e)},setSelected:function(e){return this.tree._callHook("nodeSetSelected",this,e)},setStatus:function(e,t,n){return this.tree._callHook("nodeSetStatus",this,e,t,n)},setTitle:function(e){this.title=e,this.renderTitle()},sortChildren:function(e,t){var n,r,i=this.children;if(i){if(e=e||function(e,t){var n=e.title.toLowerCase(),r=t.title.toLowerCase();return n===r?0:n>r?1:-1},i.sort(e),t)for(n=0,r=i.length;r>n;n++)i[n].children&&i[n].sortChildren(e,"$norender$");"$norender$"!==t&&this.render()}},toDict:function(t,n){var r,i,o,s={},a=this;if(e.each(S,function(e,t){(a[t]||a[t]===!1)&&(s[t]=a[t])}),e.isEmptyObject(this.data)||(s.data=e.extend({},this.data),e.isEmptyObject(s.data)&&delete s.data),n&&n(s),t&&this.hasChildren())for(s.children=[],r=0,i=this.children.length;i>r;r++)o=this.children[r],o.isStatusNode()||s.children.push(o.toDict(!0,n));return s},toggleExpanded:function(){return this.tree._callHook("nodeToggleExpanded",this)},toggleSelected:function(){return this.tree._callHook("nodeToggleSelected",this)},toString:function(){return"<FancytreeNode(#"+this.key+", '"+this.title+"')>"},visit:function(e,t){var n,r,i=!0,o=this.children;if(t===!0&&(i=e(this),i===!1||"skip"===i))return i;if(o)for(n=0,r=o.length;r>n&&(i=o[n].visit(e,!0),i!==!1);n++);return i},visitAndLoad:function(t,n,r){var i,o,s,a=this;return t&&n===!0&&(o=t(a),o===!1||"skip"===o)?r?o:u():a.children||a.lazy?(i=new e.Deferred,s=[],a.load().done(function(){for(var n=0,r=a.children.length;r>n;n++){if(o=a.children[n].visitAndLoad(t,!0,!0),o===!1){i.reject();break}"skip"!==o&&s.push(o)}e.when.apply(this,s).then(function(){i.resolve()})}),i.promise()):u()},visitParents:function(e,t){if(t&&e(this)===!1)return!1;for(var n=this.parent;n;){if(e(n)===!1)return!1;n=n.parent}return!0},warn:function(e){Array.prototype.unshift.call(arguments,this.toString()),o("warn",arguments)}},y.prototype={_makeHookContext:function(t,n,i){var o,s;return t.node!==r?(n&&t.originalEvent!==n&&e.error("invalid args"),o=t):t.tree?(s=t.tree,o={node:t,tree:s,widget:s.widget,options:s.widget.options,originalEvent:n}):t.widget?o={node:null,tree:t,widget:t.widget,options:t.widget.options,originalEvent:n}:e.error("invalid args"),i&&e.extend(o,i),o},_callHook:function(t,n,r){var i=this._makeHookContext(n),o=this[t],s=Array.prototype.slice.call(arguments,2);return e.isFunction(o)||e.error("_callHook('"+t+"') is not a function"),s.unshift(i),o.apply(this,s)},_requireExtension:function(t,n,r,o){r=!!r;var s=this._local.name,a=this.options.extensions,l=e.inArray(t,a)<e.inArray(s,a),c=n&&null==this.ext[t],u=!c&&null!=r&&r!==l;return i(s&&s!==t),c||u?(o||(c||n?(o="'"+s+"' extension requires '"+t+"'",u&&(o+=" to be registered "+(r?"before":"after")+" itself")):o="If used together, `"+t+"` must be registered "+(r?"before":"after")+" `"+s+"`"),e.error(o),!1):!0},activateKey:function(e){var t=this.getNodeByKey(e);return t?t.setActive():this.activeNode&&this.activeNode.setActive(!1),t},applyPatch:function(t){var n,r,o,s,a,l,c=t.length,u=[];for(r=0;c>r;r++)o=t[r],i(2===o.length,"patchList must be an array of length-2-arrays"),s=o[0],a=o[1],l=null===s?this.rootNode:this.getNodeByKey(s),l?(n=new e.Deferred,u.push(n),l.applyPatch(a).always(h(n,l))):this.warn("could not find node with key '"+s+"'");return e.when.apply(e,u).promise()},count:function(){return this.rootNode.countChildren()},debug:function(e){this.options.debugLevel>=2&&(Array.prototype.unshift.call(arguments,this.toString()),o("log",arguments))},findNextNode:function(t,n,r){var i=null,o=n.parent.children,s=null,a=function(e,t,n){var r,i,o=e.children,s=o.length,l=o[t];if(l&&n(l)===!1)return!1;if(l&&l.children&&l.expanded&&a(l,0,n)===!1)return!1;for(r=t+1;s>r;r++)if(a(e,r,n)===!1)return!1;return i=e.parent,i?a(i,i.children.indexOf(e)+1,n):a(e,0,n)};return t="string"==typeof t?v(t):t,n=n||this.getFirstChild(),a(n.parent,o.indexOf(n),function(r){return r===i?!1:(i=i||r,e(r.span).is(":visible")?t(r)&&(s=r,s!==n)?!1:void 0:void r.debug("quicksearch: skipping hidden node"))}),s},generateFormElements:function(t,n){var r,i=t!==!1?"ft_"+this._id+"[]":t,o=n!==!1?"ft_"+this._id+"_active":n,s="fancytree_result_"+this._id,a=e("#"+s);a.length?a.empty():a=e("<div>",{id:s}).hide().insertAfter(this.$container),i&&(r=this.getSelectedNodes(3===this.options.selectMode),e.each(r,function(t,n){a.append(e("<input>",{type:"checkbox",name:i,value:n.key,checked:!0}))})),o&&this.activeNode&&a.append(e("<input>",{type:"radio",name:o,value:this.activeNode.key,checked:!0}))},getActiveNode:function(){return this.activeNode},getFirstChild:function(){return this.rootNode.getFirstChild()},getFocusNode:function(e){return this.focusNode},getNodeByKey:function(e,t){var r,i;return!t&&(r=n.getElementById(this.options.idPrefix+e))?r.ftnode?r.ftnode:null:(t=t||this.rootNode,i=null,t.visit(function(t){return t.key===e?(i=t,!1):void 0},!0),i)},getRootNode:function(){return this.rootNode},getSelectedNodes:function(e){var t=[];return this.rootNode.visit(function(n){return n.selected&&(t.push(n),e===!0)?"skip":void 0}),t},hasFocus:function(){return!!this._hasFocus},info:function(e){this.options.debugLevel>=1&&(Array.prototype.unshift.call(arguments,this.toString()),o("info",arguments))},loadKeyPath:function(t,n,i){function o(e,t,r){n.call(y,t,"loading"),t.load().done(function(){y.loadKeyPath.call(y,d[e],n,t).always(h(r,y))}).fail(function(i){y.warn("loadKeyPath: error loading: "+e+" (parent: "+p+")"),n.call(y,t,"error"),r.reject()})}var s,a,l,c,u,d,f,p,v,g=this.options.keyPathSeparator,y=this;for(e.isArray(t)||(t=[t]),d={},l=0;l<t.length;l++)for(p=i||this.rootNode,c=t[l],c.charAt(0)===g&&(c=c.substr(1)),v=c.split(g);v.length;){if(u=v.shift(),f=p._findDirectChild(u),!f){this.warn("loadKeyPath: key not found: "+u+" (parent: "+p+")"),n.call(this,u,"error");break}if(0===v.length){n.call(this,f,"ok");break}if(f.lazy&&f.hasChildren()===r){n.call(this,f,"loaded"),d[u]?d[u].push(v.join(g)):d[u]=[v.join(g)];break}n.call(this,f,"loaded"),p=f}s=[];for(u in d)f=p._findDirectChild(u),a=new e.Deferred,s.push(a),o(u,f,a);return e.when.apply(e,s).promise()},reactivate:function(e){var t,n=this.activeNode;return n?(this.activeNode=null,t=n.setActive(),e&&n.setFocus(),t):u()},reload:function(e){return this._callHook("treeClear",this),this._callHook("treeLoad",this,e)},render:function(e,t){return this.rootNode.render(e,t)},setFocus:function(e){return this._callHook("treeSetFocus",this,e)},toDict:function(e,t){var n=this.rootNode.toDict(!0,t);return e?n:n.children},toString:function(){return"<Fancytree(#"+this._id+")>"},_triggerNodeEvent:function(e,t,n,i){var o=this._makeHookContext(t,n,i),s=this.widget._trigger(e,n,o);return s!==!1&&o.result!==r?o.result:s},_triggerTreeEvent:function(e,t,n){var i=this._makeHookContext(this,t,n),o=this.widget._trigger(e,t,i);return o!==!1&&i.result!==r?i.result:o},visit:function(e){return this.rootNode.visit(e,!1)},warn:function(e){Array.prototype.unshift.call(arguments,this.toString()),o("warn",arguments)}},e.extend(y.prototype,{nodeClick:function(e){var t,n,r=e.targetType,i=e.node;if("expander"===r)this._callHook("nodeToggleExpanded",e);else if("checkbox"===r)this._callHook("nodeToggleSelected",e),e.options.focusOnSelect&&this._callHook("nodeSetFocus",e,!0);else{if(n=!1,t=!0,i.folder)switch(e.options.clickFolderMode){case 2:n=!0,t=!1;break;case 3:t=!0,n=!0}t&&(this.nodeSetFocus(e),this._callHook("nodeSetActive",e,!0)),n&&this._callHook("nodeToggleExpanded",e)}},nodeCollapseSiblings:function(e,t){var n,r,i,o=e.node;if(o.parent)for(n=o.parent.children,r=0,i=n.length;i>r;r++)n[r]!==o&&n[r].expanded&&this._callHook("nodeSetExpanded",n[r],!1,t)},nodeDblclick:function(e){"title"===e.targetType&&4===e.options.clickFolderMode&&this._callHook("nodeToggleExpanded",e),"title"===e.targetType&&e.originalEvent.preventDefault()},nodeKeydown:function(t){var n,r,i,o=t.originalEvent,s=t.node,a=t.tree,l=t.options,c=o.which,u=String.fromCharCode(c),d=!(o.altKey||o.ctrlKey||o.metaKey||o.shiftKey),h=e(o.target),f=!0,p=!(o.ctrlKey||!l.autoActivate);if(s||(this.getFirstChild().setFocus(),s=t.node=this.focusNode,s.debug("Keydown force focus on first node")),l.quicksearch&&d&&/\w/.test(u)&&!h.is(":input:enabled"))return r=(new Date).getTime(),r-a.lastQuicksearchTime>500&&(a.lastQuicksearchTerm=""),a.lastQuicksearchTime=r,a.lastQuicksearchTerm+=u,n=a.findNextNode(a.lastQuicksearchTerm,a.getActiveNode()),n&&n.setActive(),void o.preventDefault();switch(b.keyEventToString(o)){case"+":case"=":a.nodeSetExpanded(t,!0);break;case"-":a.nodeSetExpanded(t,!1);break;case"space":l.checkbox?a.nodeToggleSelected(t):a.nodeSetActive(t,!0);break;case"enter":a.nodeSetActive(t,!0);break;case"backspace":case"left":case"right":case"up":case"down":i=s.navigate(o.which,p);break;default:f=!1}f&&o.preventDefault()},nodeLoadChildren:function(t,n){var r,o,s,a=t.tree,l=t.node;return e.isFunction(n)&&(n=n()),
+n.url&&(r=e.extend({},t.options.ajax,n),r.debugDelay?(o=r.debugDelay,e.isArray(o)&&(o=o[0]+Math.random()*(o[1]-o[0])),l.debug("nodeLoadChildren waiting debug delay "+Math.round(o)+"ms"),r.debugDelay=!1,s=e.Deferred(function(t){setTimeout(function(){e.ajax(r).done(function(){t.resolveWith(this,arguments)}).fail(function(){t.rejectWith(this,arguments)})},o)})):s=e.ajax(r),n=new e.Deferred,s.done(function(r,i,o){var s,c;if("string"==typeof r&&e.error("Ajax request returned a string (did you get the JSON dataType wrong?)."),t.options.postProcess){if(c=a._triggerNodeEvent("postProcess",t,t.originalEvent,{response:r,error:null,dataType:this.dataType}),c.error)return s=e.isPlainObject(c.error)?c.error:{message:c.error},s=a._makeHookContext(l,null,s),void n.rejectWith(this,[s]);r=e.isArray(c)?c:r}else r&&r.hasOwnProperty("d")&&t.options.enableAspx&&(r="string"==typeof r.d?e.parseJSON(r.d):r.d);n.resolveWith(this,[r])}).fail(function(e,t,r){var i=a._makeHookContext(l,null,{error:e,args:Array.prototype.slice.call(arguments),message:r,details:e.status+": "+r});n.rejectWith(this,[i])})),e.isFunction(n.promise)&&(i(!l.isLoading()),a.nodeSetStatus(t,"loading"),n.done(function(e){a.nodeSetStatus(t,"ok")}).fail(function(e){var n;n=e.node&&e.error&&e.message?e:a._makeHookContext(l,null,{error:e,args:Array.prototype.slice.call(arguments),message:e?e.message||e.toString():""}),a._triggerNodeEvent("loadError",n,null)!==!1&&a.nodeSetStatus(t,"error",n.message,n.details)})),e.when(n).done(function(t){var n;e.isPlainObject(t)&&(i(e.isArray(t.children),"source must contain (or be) an array of children"),i(l.isRoot(),"source may only be an object for root nodes"),n=t,t=t.children,delete n.children,e.extend(a.data,n)),i(e.isArray(t),"expected array of children"),l._setChildren(t),a._triggerNodeEvent("loadChildren",l)})},nodeLoadKeyPath:function(e,t){},nodeRemoveChild:function(t,n){var r,o=t.node,s=t.options,a=e.extend({},t,{node:n}),l=o.children;return 1===l.length?(i(n===l[0]),this.nodeRemoveChildren(t)):(this.activeNode&&(n===this.activeNode||this.activeNode.isDescendantOf(n))&&this.activeNode.setActive(!1),this.focusNode&&(n===this.focusNode||this.focusNode.isDescendantOf(n))&&(this.focusNode=null),this.nodeRemoveMarkup(a),this.nodeRemoveChildren(a),r=e.inArray(n,l),i(r>=0),n.visit(function(e){e.parent=null},!0),this._callHook("treeRegisterNode",this,!1,n),s.removeNode&&s.removeNode.call(t.tree,{type:"removeNode"},a),void l.splice(r,1))},nodeRemoveChildMarkup:function(t){var n=t.node;n.ul&&(n.isRoot()?e(n.ul).empty():(e(n.ul).remove(),n.ul=null),n.visit(function(e){e.li=e.ul=null}))},nodeRemoveChildren:function(t){var n,r=t.tree,i=t.node,o=i.children,s=t.options;o&&(this.activeNode&&this.activeNode.isDescendantOf(i)&&this.activeNode.setActive(!1),this.focusNode&&this.focusNode.isDescendantOf(i)&&(this.focusNode=null),this.nodeRemoveChildMarkup(t),n=e.extend({},t),i.visit(function(e){e.parent=null,r._callHook("treeRegisterNode",r,!1,e),s.removeNode&&(n.node=e,s.removeNode.call(t.tree,{type:"removeNode"},n))}),i.lazy?i.children=[]:i.children=null,this.nodeRenderStatus(t))},nodeRemoveMarkup:function(t){var n=t.node;n.li&&(e(n.li).remove(),n.li=null),this.nodeRemoveChildMarkup(t)},nodeRender:function(t,r,o,s,a){var l,c,u,d,h,f,p,v=t.node,g=t.tree,y=t.options,m=y.aria,b=!1,x=v.parent,_=!x,k=v.children;if(_||x.ul){if(i(_||x.ul,"parent UL must exist"),_||(v.li&&(r||v.li.parentNode!==v.parent.ul)&&(v.li.parentNode!==v.parent.ul&&this.debug("Unlinking "+v+" (must be child of "+v.parent+")"),this.nodeRemoveMarkup(t)),v.li?this.nodeRenderStatus(t):(b=!0,v.li=n.createElement("li"),v.li.ftnode=v,v.key&&y.generateIds&&(v.li.id=y.idPrefix+v.key),v.span=n.createElement("span"),v.span.className="fancytree-node",m&&e(v.span).attr("aria-labelledby","ftal_"+v.key),v.li.appendChild(v.span),this.nodeRenderTitle(t),y.createNode&&y.createNode.call(g,{type:"createNode"},t)),y.renderNode&&y.renderNode.call(g,{type:"renderNode"},t)),k){if(_||v.expanded||o===!0){for(v.ul||(v.ul=n.createElement("ul"),(s===!0&&!a||!v.expanded)&&(v.ul.style.display="none"),m&&e(v.ul).attr("role","group"),v.li?v.li.appendChild(v.ul):v.tree.$div.append(v.ul)),d=0,h=k.length;h>d;d++)p=e.extend({},t,{node:k[d]}),this.nodeRender(p,r,o,!1,!0);for(l=v.ul.firstChild;l;)u=l.ftnode,u&&u.parent!==v?(v.debug("_fixParent: remove missing "+u,l),f=l.nextSibling,l.parentNode.removeChild(l),l=f):l=l.nextSibling;for(l=v.ul.firstChild,d=0,h=k.length-1;h>d;d++)c=k[d],u=l.ftnode,c!==u?v.ul.insertBefore(c.li,u.li):l=l.nextSibling}}else v.ul&&(this.warn("remove child markup for "+v),this.nodeRemoveChildMarkup(t));_||b&&x.ul.appendChild(v.li)}},nodeRenderTitle:function(e,t){var n,i,o,s,a,l,c=e.node,u=e.tree,d=e.options,h=d.aria,f=c.getLevel(),p=[],v=c.data.icon;t!==r&&(c.title=t),c.span&&(f<d.minExpandLevel?(c.lazy||(c.expanded=!0),f>1&&(h?p.push("<span role='button' class='fancytree-expander fancytree-expander-fixed'></span>"):p.push("<span class='fancytree-expander fancytree-expander-fixed''></span>"))):h?p.push("<span role='button' class='fancytree-expander'></span>"):p.push("<span class='fancytree-expander'></span>"),d.checkbox&&c.hideCheckbox!==!0&&!c.isStatusNode()&&(h?p.push("<span role='checkbox' class='fancytree-checkbox'></span>"):p.push("<span class='fancytree-checkbox'></span>")),s=h?" role='img'":"",(v===!0||v!==!1&&d.icons!==!1)&&(v&&"string"==typeof v?(v="/"===v.charAt(0)?v:(d.imagePath||"")+v,p.push("<img src='"+v+"' class='fancytree-icon' alt='' />")):(i=d.iconClass&&d.iconClass.call(u,c,e)||c.data.iconclass||null,i?p.push("<span "+s+" class='fancytree-custom-icon "+i+"'></span>"):p.push("<span "+s+" class='fancytree-icon'></span>"))),o="",d.renderTitle&&(o=d.renderTitle.call(u,{type:"renderTitle"},e)||""),o||(l=c.tooltip?" title='"+b.escapeHtml(c.tooltip)+"'":"",n=h?" id='ftal_"+c.key+"'":"",s=h?" role='treeitem'":"",a=d.titlesTabbable?" tabindex='0'":"",o="<span "+s+" class='fancytree-title'"+n+l+a+">"+c.title+"</span>"),p.push(o),c.span.innerHTML=p.join(""),this.nodeRenderStatus(e))},nodeRenderStatus:function(t){var n=t.node,r=t.tree,i=t.options,o=n.hasChildren(),s=n.isLastSibling(),a=i.aria,l=e(n.span).find(".fancytree-title"),c=i._classNames,u=[],d=n[r.statusClassPropName];d&&(u.push(c.node),r.activeNode===n&&u.push(c.active),r.focusNode===n?(u.push(c.focused),a&&l.attr("aria-activedescendant",!0)):a&&l.removeAttr("aria-activedescendant"),n.expanded?(u.push(c.expanded),a&&l.attr("aria-expanded",!0)):a&&l.removeAttr("aria-expanded"),n.folder&&u.push(c.folder),o!==!1&&u.push(c.hasChildren),s&&u.push(c.lastsib),n.lazy&&null==n.children&&u.push(c.lazy),n.partsel&&u.push(c.partsel),n.unselectable&&u.push(c.unselectable),n._isLoading&&u.push(c.loading),n._error&&u.push(c.error),n.selected?(u.push(c.selected),a&&l.attr("aria-selected",!0)):a&&l.attr("aria-selected",!1),n.extraClasses&&u.push(n.extraClasses),o===!1?u.push(c.combinedExpanderPrefix+"n"+(s?"l":"")):u.push(c.combinedExpanderPrefix+(n.expanded?"e":"c")+(n.lazy&&null==n.children?"d":"")+(s?"l":"")),u.push(c.combinedIconPrefix+(n.expanded?"e":"c")+(n.folder?"f":"")),d.className=u.join(" "),n.li&&(n.li.className=s?c.lastsib:""))},nodeSetActive:function(t,n,r){r=r||{};var o,s=t.node,a=t.tree,l=t.options,c=r.noEvents===!0,h=s===a.activeNode;return n=n!==!1,h===n?u(s):n&&!c&&this._triggerNodeEvent("beforeActivate",s,t.originalEvent)===!1?d(s,["rejected"]):void(n?(a.activeNode&&(i(a.activeNode!==s,"node was active (inconsistency)"),o=e.extend({},t,{node:a.activeNode}),a.nodeSetActive(o,!1),i(null===a.activeNode,"deactivate was out of sync?")),l.activeVisible&&s.makeVisible({scrollIntoView:!1}),a.activeNode=s,a.nodeRenderStatus(t),a.nodeSetFocus(t),c||a._triggerNodeEvent("activate",s,t.originalEvent)):(i(a.activeNode===s,"node was not active (inconsistency)"),a.activeNode=null,this.nodeRenderStatus(t),c||t.tree._triggerNodeEvent("deactivate",s,t.originalEvent)))},nodeSetExpanded:function(t,n,i){i=i||{};var o,s,a,l,c,h,f=t.node,p=t.tree,v=t.options,g=i.noAnimation===!0,y=i.noEvents===!0;if(n=n!==!1,f.expanded&&n||!f.expanded&&!n)return u(f);if(n&&!f.lazy&&!f.hasChildren())return u(f);if(!n&&f.getLevel()<v.minExpandLevel)return d(f,["locked"]);if(!y&&this._triggerNodeEvent("beforeExpand",f,t.originalEvent)===!1)return d(f,["rejected"]);if(g||f.isVisible()||(g=i.noAnimation=!0),s=new e.Deferred,n&&!f.expanded&&v.autoCollapse){c=f.getParentList(!1,!0),h=v.autoCollapse;try{for(v.autoCollapse=!1,a=0,l=c.length;l>a;a++)this._callHook("nodeCollapseSiblings",c[a],i)}finally{v.autoCollapse=h}}return s.done(function(){n&&v.autoScroll&&!g?f.getLastChild().scrollIntoView(!0,{topNode:f}).always(function(){y||t.tree._triggerNodeEvent(n?"expand":"collapse",t)}):y||t.tree._triggerNodeEvent(n?"expand":"collapse",t)}),o=function(r){var i,o,s=v.toggleEffect;if(f.expanded=n,p._callHook("nodeRender",t,!1,!1,!0),f.ul)if(i="none"!==f.ul.style.display,o=!!f.expanded,i===o)f.warn("nodeSetExpanded: UL.style.display already set");else{if(s&&!g)return void e(f.ul).toggle(s.effect,s.options,s.duration,function(){r()});f.ul.style.display=f.expanded||!parent?"":"none"}r()},n&&f.lazy&&f.hasChildren()===r?f.load().done(function(){s.notifyWith&&s.notifyWith(f,["loaded"]),o(function(){s.resolveWith(f)})}).fail(function(e){o(function(){s.rejectWith(f,["load failed ("+e+")"])})}):o(function(){s.resolveWith(f)}),s.promise()},nodeSetFocus:function(t,n){var r,i=t.tree,o=t.node;if(n=n!==!1,i.focusNode){if(i.focusNode===o&&n)return;r=e.extend({},t,{node:i.focusNode}),i.focusNode=null,this._triggerNodeEvent("blur",r),this._callHook("nodeRenderStatus",r)}n&&(this.hasFocus()||(o.debug("nodeSetFocus: forcing container focus"),this._callHook("treeSetFocus",t,!0,!0)),o.makeVisible({scrollIntoView:!1}),i.focusNode=o,this._triggerNodeEvent("focus",t),t.options.autoScroll&&o.scrollIntoView(),this._callHook("nodeRenderStatus",t))},nodeSetSelected:function(e,t){var n=e.node,r=e.tree,i=e.options;if(t=t!==!1,n.debug("nodeSetSelected("+t+")",e),!n.unselectable){if(n.selected&&t||!n.selected&&!t)return!!n.selected;if(this._triggerNodeEvent("beforeSelect",n,e.originalEvent)===!1)return!!n.selected;t&&1===i.selectMode?r.lastSelectedNode&&r.lastSelectedNode.setSelected(!1):3===i.selectMode&&(n.selected=t,n.fixSelection3AfterClick()),n.selected=t,this.nodeRenderStatus(e),r.lastSelectedNode=t?n:null,r._triggerNodeEvent("select",e)}},nodeSetStatus:function(t,n,r,i){function o(){var e=a.children?a.children[0]:null;if(e&&e.isStatusNode()){try{a.ul&&(a.ul.removeChild(e.li),e.li=null)}catch(t){}1===a.children.length?a.children=[]:a.children.shift()}}function s(t,n){var r=a.children?a.children[0]:null;return r&&r.isStatusNode()?(e.extend(r,t),l._callHook("nodeRenderTitle",r)):(t.key="_statusNode",a._setChildren([t]),a.children[0].statusNodeType=n,l.render()),a.children[0]}var a=t.node,l=t.tree;switch(n){case"ok":o(),a._isLoading=!1,a._error=null,a.renderStatus();break;case"loading":a.parent||s({title:l.options.strings.loading+(r?" ("+r+") ":""),tooltip:i,extraClasses:"fancytree-statusnode-wait"},n),a._isLoading=!0,a._error=null,a.renderStatus();break;case"error":s({title:l.options.strings.loadError+(r?" ("+r+") ":""),tooltip:i,extraClasses:"fancytree-statusnode-error"},n),a._isLoading=!1,a._error={message:r,details:i},a.renderStatus();break;default:e.error("invalid node status "+n)}},nodeToggleExpanded:function(e){return this.nodeSetExpanded(e,!e.node.expanded)},nodeToggleSelected:function(e){return this.nodeSetSelected(e,!e.node.selected)},treeClear:function(e){var t=e.tree;t.activeNode=null,t.focusNode=null,t.$div.find(">ul.fancytree-container").empty(),t.rootNode.children=null},treeCreate:function(e){},treeDestroy:function(e){},treeInit:function(e){this.treeLoad(e)},treeLoad:function(t,n){var r,i,o,s=t.tree,a=t.widget.element,l=e.extend({},t,{node:this.rootNode});if(s.rootNode.children&&this.treeClear(t),n=n||this.options.source)"string"==typeof n&&e.error("Not implemented");else switch(r=a.data("type")||"html"){case"html":i=a.find(">ul:first"),i.addClass("ui-fancytree-source ui-helper-hidden"),n=e.ui.fancytree.parseHtml(i),this.data=e.extend(this.data,f(i));break;case"json":n=e.parseJSON(a.text()),n.children&&(n.title&&(s.title=n.title),n=n.children);break;default:e.error("Invalid data-type: "+r)}return o=this.nodeLoadChildren(l,n).done(function(){s.render(),3===t.options.selectMode&&s.rootNode.fixSelection3FromEndNodes(),s._triggerTreeEvent("init",null,{status:!0})}).fail(function(){s.render(),s._triggerTreeEvent("init",null,{status:!1})})},treeRegisterNode:function(e,t,n){},treeSetFocus:function(e,t,n){t=t!==!1,t!==this.hasFocus()&&(this._hasFocus=t,this.$container.toggleClass("fancytree-treefocus",t),this._triggerTreeEvent(t?"focusTree":"blurTree"))}}),e.widget("ui.fancytree",{options:{activeVisible:!0,ajax:{type:"GET",cache:!1,dataType:"json"},aria:!1,autoActivate:!0,autoCollapse:!1,autoScroll:!1,checkbox:!1,clickFolderMode:4,debugLevel:null,disabled:!1,enableAspx:!0,extensions:[],toggleEffect:{effect:"blind",options:{direction:"vertical",scale:"box"},duration:200},generateIds:!1,icons:!0,idPrefix:"ft_",focusOnSelect:!1,keyboard:!0,keyPathSeparator:"/",minExpandLevel:1,quicksearch:!1,scrollOfs:{top:0,bottom:0},scrollParent:null,selectMode:2,strings:{loading:"Loading&#8230;",loadError:"Load error!"},tabbable:!0,titlesTabbable:!1,_classNames:{node:"fancytree-node",folder:"fancytree-folder",combinedExpanderPrefix:"fancytree-exp-",combinedIconPrefix:"fancytree-ico-",hasChildren:"fancytree-has-children",active:"fancytree-active",selected:"fancytree-selected",expanded:"fancytree-expanded",lazy:"fancytree-lazy",focused:"fancytree-focused",partsel:"fancytree-partsel",unselectable:"fancytree-unselectable",lastsib:"fancytree-lastsib",loading:"fancytree-loading",error:"fancytree-error"},lazyLoad:null,postProcess:null},_create:function(){this.tree=new y(this),this.$source=this.source||"json"===this.element.data("type")?this.element:this.element.find(">ul:first");var t,n,o,s=this.options.extensions,a=this.tree;for(o=0;o<s.length;o++)n=s[o],t=e.ui.fancytree._extensions[n],t||e.error("Could not apply extension '"+n+"' (it is not registered, did you forget to include it?)"),this.tree.options[n]=e.extend(!0,{},t.options,this.tree.options[n]),i(this.tree.ext[n]===r,"Extension name must not exist as Fancytree.ext attribute: '"+n+"'"),this.tree.ext[n]={},c(this.tree,a,t,n),a=t;this.tree._callHook("treeCreate",this.tree)},_init:function(){this.tree._callHook("treeInit",this.tree),this._bind()},_setOption:function(t,n){var r=!0,i=!1;switch(t){case"aria":case"checkbox":case"icons":case"minExpandLevel":case"tabbable":this.tree._callHook("treeCreate",this.tree),i=!0;break;case"source":r=!1,this.tree._callHook("treeLoad",this.tree,n)}this.tree.debug("set option "+t+"="+n+" <"+typeof n+">"),r&&e.Widget.prototype._setOption.apply(this,arguments),i&&this.tree.render(!0,!1)},destroy:function(){this._unbind(),this.tree._callHook("treeDestroy",this.tree),this.tree.$div.find(">ul.fancytree-container").remove(),this.$source&&this.$source.removeClass("ui-helper-hidden"),e.Widget.prototype.destroy.call(this)},_unbind:function(){var t=this.tree._ns;this.element.unbind(t),this.tree.$container.unbind(t),e(n).unbind(t)},_bind:function(){var e=this,t=this.options,n=this.tree,r=n._ns;this._unbind(),n.$container.on("focusin"+r+" focusout"+r,function(e){var t=b.getNode(e),r="focusin"===e.type;t?n._callHook("nodeSetFocus",t,r):n._callHook("treeSetFocus",n,r)}).on("selectstart"+r,"span.fancytree-title",function(e){e.preventDefault()}).on("keydown"+r,function(e){if(t.disabled||t.keyboard===!1)return!0;var r,i=n.focusNode,o=n._makeHookContext(i||n,e),s=n.phase;try{return n.phase="userEvent",r=i?n._triggerNodeEvent("keydown",i,e):n._triggerTreeEvent("keydown",e),"preventNav"===r?r=!0:r!==!1&&(r=n._callHook("nodeKeydown",o)),r}finally{n.phase=s}}).on("click"+r+" dblclick"+r,function(n){if(t.disabled)return!0;var r,i=b.getEventTarget(n),o=i.node,s=e.tree,a=s.phase;if(!o)return!0;r=s._makeHookContext(o,n);try{switch(s.phase="userEvent",n.type){case"click":return r.targetType=i.type,s._triggerNodeEvent("click",r,n)===!1?!1:s._callHook("nodeClick",r);case"dblclick":return r.targetType=i.type,s._triggerNodeEvent("dblclick",r,n)===!1?!1:s._callHook("nodeDblclick",r)}}finally{s.phase=a}})},getActiveNode:function(){return this.tree.activeNode},getNodeByKey:function(e){return this.tree.getNodeByKey(e)},getRootNode:function(){return this.tree.rootNode},getTree:function(){return this.tree}}),b=e.ui.fancytree,e.extend(e.ui.fancytree,{version:"2.7.0",buildType:"production",debugLevel:1,_nextId:1,_nextNodeKey:1,_extensions:{},_FancytreeClass:y,_FancytreeNodeClass:g,jquerySupports:{positionMyOfs:a(e.ui.version,1,9)},assert:function(e,t){return i(e,t)},debounce:function(e,t,n,r){var i;return 3===arguments.length&&"boolean"!=typeof n&&(r=n,n=!1),function(){var o=arguments;r=r||this,n&&!i&&t.apply(r,o),clearTimeout(i),i=setTimeout(function(){n||t.apply(r,o),i=null},e)}},debug:function(t){e.ui.fancytree.debugLevel>=2&&o("log",arguments)},error:function(e){o("error",arguments)},escapeHtml:function(e){return(""+e).replace(/[&<>"'\/]/g,function(e){return x[e]})},getEventTargetType:function(e){return this.getEventTarget(e).type},getEventTarget:function(t){var n=t&&t.target?t.target.className:"",i={node:this.getNode(t.target),type:r};return/\bfancytree-title\b/.test(n)?i.type="title":/\bfancytree-expander\b/.test(n)?i.type=i.node.hasChildren()===!1?"prefix":"expander":/\bfancytree-checkbox\b/.test(n)||/\bfancytree-radio\b/.test(n)?i.type="checkbox":/\bfancytree-icon\b/.test(n)?i.type="icon":/\bfancytree-node\b/.test(n)?i.type="title":t&&t.target&&e(t.target).closest(".fancytree-title").length&&(i.type="title"),i},getNode:function(e){if(e instanceof g)return e;for(e.selector!==r?e=e[0]:e.originalEvent!==r&&(e=e.target);e;){if(e.ftnode)return e.ftnode;e=e.parentNode}return null},info:function(t){e.ui.fancytree.debugLevel>=1&&o("info",arguments)},keyEventToString:function(e){var t=e.which,n=[];return e.altKey&&n.push("alt"),e.ctrlKey&&n.push("ctrl"),e.metaKey&&n.push("meta"),e.shiftKey&&n.push("shift"),_[t]||n.push(k[t]||String.fromCharCode(t).toLowerCase()),n.join("+")},parseHtml:function(t){var n,i,o,s,a,l,c,u,d=t.find(">li"),h=[];return d.each(function(){var d,p=e(this),v=p.find(">span:first",this),g=v.length?null:p.find(">a:first"),y={tooltip:null,data:{}};for(v.length?y.title=v.html():g&&g.length?(y.title=g.html(),y.data.href=g.attr("href"),y.data.target=g.attr("target"),y.tooltip=g.attr("title")):(y.title=p.html(),s=y.title.search(/<ul/i),s>=0&&(y.title=y.title.substring(0,s))),y.title=e.trim(y.title),i=0,o=N.length;o>i;i++)y[N[i]]=r;for(c=this.className.split(" "),n=[],i=0,o=c.length;o>i;i++)u=c[i],w[u]?y[u]=!0:n.push(u);if(y.extraClasses=n.join(" "),a=p.attr("title"),a&&(y.tooltip=a),a=p.attr("id"),a&&(y.key=a),d=f(p),d&&!e.isEmptyObject(d)){for(i=0,o=S.length;o>i;i++)a=S[i],l=d[a],null!=l&&(delete d[a],y[a]=l);e.extend(y.data,d)}t=p.find(">ul:first"),t.length?y.children=e.ui.fancytree.parseHtml(t):y.children=y.lazy?r:null,h.push(y)}),h},registerExtension:function(t){i(null!=t.name,"extensions must have a `name` property."),i(null!=t.version,"extensions must have a `version` property."),e.ui.fancytree._extensions[t.name]=t},unescapeHtml:function(e){var t=n.createElement("div");return t.innerHTML=e,0===t.childNodes.length?"":t.childNodes[0].nodeValue},warn:function(e){o("warn",arguments)}})}(jQuery,window,document)},function(e,t,n){"use strict";function r(e){return e&&e.__esModule?e:{"default":e}}function i(e){return e&&"undefined"!=typeof Symbol&&e.constructor===Symbol?"symbol":typeof e}function o(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}function s(e,t){var n="?"===e.substr(0,1),r=e.substr(1).split("."),i=t;for(prop in r)if(i=i[r[prop]],void 0===i){if(n)break;throw'Attribute "'+r[prop]+'" does not exists'}return i}var a=function(){function e(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}return function(t,n,r){return n&&e(t.prototype,n),r&&e(t,r),t}}();Object.defineProperty(t,"__esModule",{value:!0}),t.FancytreeAdapter=void 0;var l=n(31),c=r(l);n(30),n(28),n(54),n(53);var u=new c["default"];t.FancytreeAdapter=function(){function e(t){if(o(this,e),!window.jQuery||!jQuery.fn.fancytree)throw"The FancytreeAdapter requires both jQuery and the FancyTree library.";if(!t.request)throw"The FancytreeAdapter requires a request option.";this.requestData=t.request,this.rootNode=t.root_node||"/",this.useCache=void 0===t.use_cache?!0:t.use_cache,this.actions=new c["default"],this.tree=null,this.$tree=null}return a(e,[{key:"bindToElement",value:function(e){if(this.$tree)throw"Cannot bind to multiple elements.";if(!e instanceof jQuery)throw"FancytreeAdapter can only be adapted to a jQuery object.";this.$tree=e;var t=this.actions,n=(this.requestNode,function a(e){if(0!==e.length){var n={title:e.label,key:e.node_name,children:[],actions:{}};for(var r in t)if(t.has(r)){var o=t.get(r),l=o.url;"object"==i(o.url)&&o.url.hasOwnProperty("data")&&(l=s(o.url.data,e)),void 0!==l&&(n.actions[r]={label:r,iconClass:o.icon,url:l})}var c=0;for(name in e.children)if(e.children.hasOwnProperty(name)){var u=a(e.children[name]);u&&n.children.push(u),c++}return c&&(n.folder=!0,n.lazy=!0),n}}),r=this.requestData,o=this.useCache;this.$tree.fancytree({source:o&&u.has(this.rootNode)?u.get(this.rootNode):r.load(this.rootNode),lazyLoad:function(e,t){var n=t.node.getKeyPath();if(o&&u.has(n))t.result=u.get(n);else{var i=r.load(n);Array.isArray(i)?t.result=i:t.result=jQuery.extend({data:{}},i)}},postProcess:function(e,t){null==t.error?(t.result=n(t.response).children,1==t.result.length&&(t.result[0].expanded=!0),o&&u.set(t.node.getKeyPath(),t.result)):t.result={error:"An error occured while retrieving the nodes: "+t.error}},activeVisible:!0}),this.actions&&this.$tree.cmfContextMenu({delegate:"span.fancytree-title",wrapperTemplate:'<ul class="dropdown-menu" style="display:block;"></ul>',actionTemplate:'<li role="presentation"><a role="menuitem" href="{{ url }}"><i class="{{ iconClass }}"></i> {{ label }}</li>',actions:function(e){return jQuery.ui.fancytree.getNode(e).data.actions}}),this.tree=this.$tree.fancytree("getTree")}},{key:"bindToInput",value:function(e){this.$tree.fancytree("option","activate",function(t,n){e.val(n.node.getKeyPath())});var t=this.tree,n=function(e){t.loadKeyPath(e,function(e,t){"ok"==t&&(e.setExpanded(),e.setActive())})};this.$tree.bind("fancytreeinit",function(t,r){n(e.val())}),e.on("change",function(e){n($(this).val())})}},{key:"addAction",value:function(e,t,n){this.actions.set(e,{url:t,icon:n})}}],[{key:"_resetCache",value:function(){u.clear()}}]),e}()},function(e,t,n){"use strict";function r(e){return e&&e.__esModule?e:{"default":e}}var i=n(27),o=r(i);o["default"].fn.cmfContextMenu=function(e){var t,e=o["default"].extend({delegate:null,actions:[],select:function(e,t){},wrapperTemplate:'<ul id="cmf-context-menu"></ul>',actionTemplate:'<li><i class="{{ iconClass }}"></i> {{ label }}</li>'},e),n=(0,o["default"])("body");(0,o["default"])(this).on("contextmenu",e.delegate,function(r){r.preventDefault();var i=(0,o["default"])(this);t&&t.remove();var s=e.actions;if("function"==typeof s&&(s=s(i)),!1!==s&&!o["default"].isEmptyObject(s)){t=function(){var t=(0,o["default"])(e.wrapperTemplate),n=t.is("ul")?t:t.find("ul");for(var r in s){var i=s[r],a=(0,o["default"])(function(){var t=e.actionTemplate;for(var n in i)i.hasOwnProperty(n)&&(t=t.replace("{{ "+n+" }}",i[n]));return t}());a.data("cmd",r),n.append(a)}return t}(),t.css({top:r.pageY,left:r.pageX}),n.append(t);var a=e.select;t.on("click","li",function(e){e.stopPropagation(),a(i,e)})}}),(0,o["default"])("html").on("click",function(e){t&&t.remove()})}},function(e,t,n){n(50),n(51),n(52),n(49),e.exports=n(7).Map},function(e,t){e.exports=function(e){if("function"!=typeof e)throw TypeError(e+" is not a function!");return e}},function(e,t,n){var r=n(1)("unscopables"),i=Array.prototype;void 0==i[r]&&n(3)(i,r,{}),e.exports=function(e){i[r][e]=!0}},function(e,t,n){"use strict";var r=n(4),i=n(3),o=n(24),s=n(8),a=n(25),l=n(9),c=n(21),u=n(13),d=n(22),h=n(15)("id"),f=n(11),p=n(12),v=n(42),g=n(10),y=Object.isExtensible||p,m=g?"_s":"size",b=0,x=function(e,t){if(!p(e))return"symbol"==typeof e?e:("string"==typeof e?"S":"P")+e;if(!f(e,h)){if(!y(e))return"F";if(!t)return"E";i(e,h,++b)}return"O"+e[h]},_=function(e,t){var n,r=x(t);if("F"!==r)return e._i[r];for(n=e._f;n;n=n.n)if(n.k==t)return n};e.exports={getConstructor:function(e,t,n,i){var u=e(function(e,o){a(e,u,t),e._i=r.create(null),e._f=void 0,e._l=void 0,e[m]=0,void 0!=o&&c(o,n,e[i],e)});return o(u.prototype,{clear:function(){for(var e=this,t=e._i,n=e._f;n;n=n.n)n.r=!0,n.p&&(n.p=n.p.n=void 0),delete t[n.i];e._f=e._l=void 0,e[m]=0},"delete":function(e){var t=this,n=_(t,e);if(n){var r=n.n,i=n.p;delete t._i[n.i],n.r=!0,i&&(i.n=r),r&&(r.p=i),t._f==n&&(t._f=r),t._l==n&&(t._l=i),t[m]--}return!!n},forEach:function(e){for(var t,n=s(e,arguments.length>1?arguments[1]:void 0,3);t=t?t.n:this._f;)for(n(t.v,t.k,this);t&&t.r;)t=t.p},has:function(e){return!!_(this,e)}}),g&&r.setDesc(u.prototype,"size",{get:function(){return l(this[m])}}),u},def:function(e,t,n){var r,i,o=_(e,t);return o?o.v=n:(e._l=o={i:i=x(t,!0),k:t,v:n,p:r=e._l,n:void 0,r:!1},e._f||(e._f=o),r&&(r.n=o),e[m]++,"F"!==i&&(e._i[i]=o)),e},getEntry:_,setStrong:function(e,t,n){u(e,t,function(e,t){this._t=e,this._k=t,this._l=void 0},function(){for(var e=this,t=e._k,n=e._l;n&&n.r;)n=n.p;return e._t&&(e._l=n=n?n.n:e._t._f)?"keys"==t?d(0,n.k):"values"==t?d(0,n.v):d(0,[n.k,n.v]):(e._t=void 0,d(1))},n?"entries":"values",!n,!0),v(t)}}},function(e,t,n){"use strict";var r=n(2),i=n(19),o=n(6),s=n(24),a=n(21),l=n(25),c=n(12),u=n(20),d=n(40),h=n(14);e.exports=function(e,t,n,f,p,v){var g=r[e],y=g,m=p?"set":"add",b=y&&y.prototype,x={},_=function(e){var t=b[e];o(b,e,"delete"==e?function(e){return v&&!c(e)?!1:t.call(this,0===e?0:e)}:"has"==e?function(e){return v&&!c(e)?!1:t.call(this,0===e?0:e)}:"get"==e?function(e){return v&&!c(e)?void 0:t.call(this,0===e?0:e)}:"add"==e?function(e){return t.call(this,0===e?0:e),this}:function(e,n){return t.call(this,0===e?0:e,n),this})};if("function"==typeof y&&(v||b.forEach&&!u(function(){(new y).entries().next()}))){var k,N=new y,w=N[m](v?{}:-0,1)!=N,S=u(function(){N.has(1)}),C=d(function(e){new y(e)});C||(y=t(function(t,n){l(t,y,e);var r=new g;return void 0!=n&&a(n,p,r[m],r),r}),y.prototype=b,b.constructor=y),v||N.forEach(function(e,t){k=1/t===-(1/0)}),(S||k)&&(_("delete"),_("has"),p&&_("get")),(k||w)&&_(m),v&&b.clear&&delete b.clear}else y=f.getConstructor(t,e,p,m),s(y.prototype,n);return h(y,e),x[e]=y,i(i.G+i.W+i.F*(y!=g),x),v||f.setStrong(y,e,p),y}},function(e,t,n){var r=n(18);e.exports=Object("z").propertyIsEnumerable(0)?Object:function(e){return"String"==r(e)?e.split(""):Object(e)}},function(e,t,n){var r=n(5),i=n(1)("iterator"),o=Array.prototype;e.exports=function(e){return void 0!==e&&(r.Array===e||o[i]===e)}},function(e,t,n){var r=n(16);e.exports=function(e,t,n,i){try{return i?t(r(n)[0],n[1]):t(n)}catch(o){var s=e["return"];throw void 0!==s&&r(s.call(e)),o}}},function(e,t,n){"use strict";var r=n(4),i=n(23),o=n(14),s={};n(3)(s,n(1)("iterator"),function(){return this}),e.exports=function(e,t,n){e.prototype=r.create(s,{next:i(1,n)}),o(e,t+" Iterator")}},function(e,t,n){var r=n(1)("iterator"),i=!1;try{var o=[7][r]();o["return"]=function(){i=!0},Array.from(o,function(){throw 2})}catch(s){}e.exports=function(e,t){if(!t&&!i)return!1;var n=!1;try{var o=[7],s=o[r]();s.next=function(){n=!0},o[r]=function(){return s},e(o)}catch(a){}return n}},function(e,t){e.exports=!1},function(e,t,n){"use strict";var r=n(2),i=n(4),o=n(10),s=n(1)("species");e.exports=function(e){var t=r[e];o&&t&&!t[s]&&i.setDesc(t,s,{configurable:!0,get:function(){return this}})}},function(e,t,n){var r=n(2),i="__core-js_shared__",o=r[i]||(r[i]={});e.exports=function(e){return o[e]||(o[e]={})}},function(e,t,n){var r=n(26),i=n(9);e.exports=function(e){return function(t,n){var o,s,a=String(i(t)),l=r(n),c=a.length;return 0>l||l>=c?e?"":void 0:(o=a.charCodeAt(l),55296>o||o>56319||l+1===c||(s=a.charCodeAt(l+1))<56320||s>57343?e?a.charAt(l):o:e?a.slice(l,l+2):(o-55296<<10)+(s-56320)+65536)}}},function(e,t,n){var r=n(36),i=n(9);e.exports=function(e){return r(i(e))}},function(e,t,n){var r=n(26),i=Math.min;e.exports=function(e){return e>0?i(r(e),9007199254740991):0}},function(e,t,n){var r=n(17),i=n(1)("iterator"),o=n(5);e.exports=n(7).getIteratorMethod=function(e){return void 0!=e?e[i]||e["@@iterator"]||o[r(e)]:void 0}},function(e,t,n){"use strict";var r=n(33),i=n(22),o=n(5),s=n(45);e.exports=n(13)(Array,"Array",function(e,t){this._t=s(e),this._i=0,this._k=t},function(){var e=this._t,t=this._k,n=this._i++;return!e||n>=e.length?(this._t=void 0,i(1)):"keys"==t?i(0,n):"values"==t?i(0,e[n]):i(0,[n,e[n]])},"values"),o.Arguments=o.Array,r("keys"),r("values"),r("entries")},function(e,t,n){"use strict";var r=n(34);n(35)("Map",function(e){return function(){return e(this,arguments.length>0?arguments[0]:void 0)}},{get:function(e){var t=r.getEntry(this,e);return t&&t.v},set:function(e,t){return r.def(this,0===e?0:e,t)}},r,!0)},function(e,t,n){"use strict";var r=n(17),i={};i[n(1)("toStringTag")]="z",i+""!="[object z]"&&n(6)(Object.prototype,"toString",function(){return"[object "+r(this)+"]"},!0)},function(e,t,n){"use strict";var r=n(44)(!0);n(13)(String,"String",function(e){this._t=String(e),this._i=0},function(){var e,t=this._t,n=this._i;return n>=t.length?{value:void 0,done:!0}:(e=r(t,n),this._i+=e.length,{value:e,done:!1})})},function(e,t,n){n(48);var r=n(2),i=n(3),o=n(5),s=n(1)("iterator"),a=r.NodeList,l=r.HTMLCollection,c=a&&a.prototype,u=l&&l.prototype,d=o.NodeList=o.HTMLCollection=o.Array;c&&!c[s]&&i(c,s,d),u&&!u[s]&&i(u,s,d)},function(e,t){},function(e,t){}]);
