+++ Resources/views/Base/scripts.html.twig
@@ -1,14 +1,10 @@
-{% set treeAdapter = treeAdapter|default('fancytree') %}
+{% set treeAdapter = treeAdapter|default('fancytree')|lower %}
 
 {% if treeAdapter == 'fancytree' %}
-<script>window.jQuery || document.write(unescape('%3Cscript src="{{ asset(assetsBasePath ~ '/jquery/dist/jquery.min.js') }}"%3E%3C/script%3E'))</script>
-<script>jQuery.ui || document.write(unescape('%3Cscript src="{{ asset(assetsBasePath ~ '/jquery-ui/jquery-ui.min.js') }}"%3E%3C/script%3E'))</script>
-<script src="{{ asset(assetsBasePath ~ '/fancytree/dist/src/jquery.fancytree.js') }}" type="text/javascript"></script>
+<script>window.jQuery || document.write('<script src="{{ asset(assetsBasePath ~ '/jquery/dist/jquery.min.js') }}"><\/script>')</script>
+<script>window.jQuery.ui || document.write('<script src="{{ asset(assetsBasePath ~ '/jquery-ui/jquery-ui.min.js') }}"><\/script>')</script>
 {% endif %}
 
-<script src="{{ asset('bundles/cmftreebrowser/js/adapter/' ~ treeAdapter ~ '.js') }}"></script>
-<script src="{{ asset('bundles/cmftreebrowser/js/jquery.cmf_context_menu.js') }}"></script>
-<script src="{{ asset('bundles/cmftreebrowser/js/jquery.cmf_tree.js') }}"></script>
+<script src="{{ asset('bundles/cmftreebrowser/js/cmf_tree_browser.' ~ treeAdapter ~ '.js') }}" type="text/javascript"></script>
 
-<link rel="stylesheet" href="{{ asset(assetsBasePath ~ '/fancytree/dist/skin-win8/ui.fancytree.min.css') }}" media="all"/>
-<link rel="stylesheet" href="{{ asset('bundles/cmftreebrowser/css/fontawesome-style.css') }}" media="all"/>
+<link rel="stylesheet" href="{{ asset('bundles/cmftreebrowser/css/cmf_tree_browser.' ~ treeAdapter ~ '.css') }}" media="all"/>
