+++ Tests/tests.cmf_tree_browser.js
@@ -1,0 +1,3 @@
+var context = require.context('./js', true, /.+Spec\.js$/);
+context.keys().forEach(context);
+module.exports = context;
