+++ package.json
@@ -1,29 +1,35 @@
 {
   "name": "symfony-cmf-tree-browser-bundle",
-  "dependencies": {
-    "event-stream": "^3.3.0",
-    "gulp": "^3.8.11",
-    "gulp-concat": "^2.5.2",
-    "gulp-bower": "^0.0.10",
-    "gulp-uglify": "^1.1.0"
-  },
+  "license": "MIT",
   "devDependencies": {
+    "babel-core": "^6.3.17",
+    "babel-loader": "^6.2.0",
+    "babel-preset-es2015": "^6.3.13",
+    "core-js": "^1.2.6",
+    "css-loader": "^0.23.0",
+    "extract-text-webpack-plugin": "^0.9.1",
+    "file-loader": "^0.8.5",
     "jasmine-ajax": "^3.1.0",
     "jasmine-core": "^2.2.0",
     "jasmine-jquery": "^2.0.6",
-    "karma": "^0.12.31",
+    "karma": "^0.13.0",
     "karma-chrome-launcher": "^0.1.7",
     "karma-firefox-launcher": "^0.1.4",
     "karma-ie-launcher": "^0.1.5",
     "karma-jasmine": "^0.3.5",
-    "karma-phantomjs-launcher": "^0.1.4"
+    "karma-phantomjs-launcher": "^0.1.4",
+    "karma-sourcemap-loader": "^0.3.6",
+    "karma-webpack": "^1.7.0",
+    "style-loader": "^0.13.0",
+    "webpack": "^1.12.9"
   },
   "scripts": {
-    "test": "\"node_modules/karma/bin/karma\" start --single-run --browsers PhantomJS"
+    "test": "\"node_modules/karma/bin/karma\" start --single-run --browsers PhantomJS",
+    "build": "sh _build.sh"
   },
   "repository": {
-     "type": "git",
-     "url": "https://github.com/symfony-cmf/TreeBrowserBundle.git"
-   },
+    "type": "git",
+    "url": "https://github.com/symfony-cmf/TreeBrowserBundle.git"
+  },
   "homepage": "https://github.com/symfony-cmf/TreeBrowserBundle"
 }
